Implement by-pid and by-name plists, refactor
