+++ memflow-framework/Cargo.toml
@@ -12,3 +12,4 @@ cglue = { version = ">=0.2.5", default-features = false, features = ["layout_che
 filer = { version = "0.1", path = "../filer" }
 once_cell = "1.9"
 num = "0.4"
+dashmap = "5"
